Add support for fractional hours.

+++ DuffTimeCard/DuffTimeCard/Task.h
@@ -11,12 +11,12 @@
 @interface Task : NSObject
 
 @property (strong, nonatomic) NSString *name;
-@property (nonatomic) int hours;
+@property (nonatomic) double hours;
 @property (nonatomic) int projectIndex;
 @property (strong, nonatomic) NSString *notes;
 @property (strong, nonatomic) NSString *date;
 
-- (Task *)initWithName:(NSString *)name hours:(int)hours projectIndex:(int)projectIndex notes:(NSString *)notes date:(NSString *)date;
+- (Task *)initWithName:(NSString *)name hours:(double)hours projectIndex:(int)projectIndex notes:(NSString *)notes date:(NSString *)date;
 - (NSDictionary *)createJSONObjectFromTask;
 
 @end
